vec ops size i32->u64

+++ icicle_v3/backend/cpu/src/field/cpu_vec_ops.cpp
@@ -11,9 +11,9 @@ using namespace icicle;
 /*********************************** ADD ***********************************/
 template <typename T>
 eIcicleError
-cpu_vector_add(const Device& device, const T* vec_a, const T* vec_b, int n, const VecOpsConfig& config, T* output)
+cpu_vector_add(const Device& device, const T* vec_a, const T* vec_b, uint64_t n, const VecOpsConfig& config, T* output)
 {
-  for (int i = 0; i < n; ++i) {
+  for (uint64_t i = 0; i < n; ++i) {
     output[i] = vec_a[i] + vec_b[i];
   }
   return eIcicleError::SUCCESS;
@@ -24,9 +24,9 @@ REGISTER_VECTOR_ADD_BACKEND("CPU", cpu_vector_add<scalar_t>);
 /*********************************** SUB ***********************************/
 template <typename T>
 eIcicleError
-cpu_vector_sub(const Device& device, const T* vec_a, const T* vec_b, int n, const VecOpsConfig& config, T* output)
+cpu_vector_sub(const Device& device, const T* vec_a, const T* vec_b, uint64_t n, const VecOpsConfig& config, T* output)
 {
-  for (int i = 0; i < n; ++i) {
+  for (uint64_t i = 0; i < n; ++i) {
     output[i] = vec_a[i] - vec_b[i];
   }
   return eIcicleError::SUCCESS;
@@ -37,9 +37,9 @@ REGISTER_VECTOR_SUB_BACKEND("CPU", cpu_vector_sub<scalar_t>);
 /*********************************** MUL ***********************************/
 template <typename T>
 eIcicleError
-cpu_vector_mul(const Device& device, const T* vec_a, const T* vec_b, int n, const VecOpsConfig& config, T* output)
+cpu_vector_mul(const Device& device, const T* vec_a, const T* vec_b, uint64_t n, const VecOpsConfig& config, T* output)
 {
-  for (int i = 0; i < n; ++i) {
+  for (uint64_t i = 0; i < n; ++i) {
     output[i] = vec_a[i] * vec_b[i];
   }
   return eIcicleError::SUCCESS;
